backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
pymysql==1.1.1
cryptography==42.0.8
pydantic==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
redis==5.0.4
anyio==4.4.0
email-validator==2.1.1
openpyxl==3.1.2
pandas==2.2.2
xlsxwriter==3.2.0
jinja2==3.1.4
aiosmtplib==3.0.1
slowapi==0.1.9
bandit==1.7.8
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
factory-boy==3.3.0
